+++ MulliganMadness/Cards/CardRegistration.cs
@@ -1,3 +1,5 @@
+using System;
+using MulliganMadness.Utils;
 using UnboundLib.Cards;
 
 namespace MulliganMadness.Cards
@@ -6,8 +8,19 @@ namespace MulliganMadness.Cards
     {
         internal static void RegisterAll()
         {
-            CustomCard.BuildCard<NestEgg>(info => NestEgg.Card = info);
-            CustomCard.BuildCard<ReturnToSender>(info => ReturnToSender.Card = info);
+            Bind<NestEgg>(info => NestEgg.Card = info);
+            Bind<ReturnToSender>(info => ReturnToSender.Card = info);
+        }
+
+        private static void Bind<T>(Action<CardInfo> setStatic) where T : CustomCard
+        {
+            CustomCard.BuildCard<T>(info =>
+            {
+                setStatic(info);
+                // Unbound sets cardArt and cardName after SetupCard. The callback
+                // is the first moment both exist on the prefab.
+                CardArtFactory.TryAssignSprite(info);
+            });
         }
     }
 }